Adding styles view should now be trivial
